+++ crates/app/src/cli/commands/mod.rs
@@ -940,7 +940,7 @@ pub async fn cli() -> Result<(), Box<dyn Error>> {
             println!("{}", "success".success());
             println!("{}", "warning".warning());
             eprintln!("{}", "error".error());
-            println!("{}", "https://github.com/oxy-hq/oxy/".secondary());
+            println!("{}", "https://github.com/oxy-hq/oxygen/".secondary());
             println!("{}", "-region".tertiary());
             println!("{}", "Viewing repository".info());
             println!("{}", "text".text());