+++ crates/app/src/main.rs
@@ -166,7 +166,7 @@ fn main() {
         setup_panic!(
             Metadata::new(env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"))
                 .authors("Robert Yi <robert@oxy.tech>") // temporarily using Robert email here, TODO: replace by support email
-                .homepage("github.com/oxy-hq/oxy")
+                .homepage("github.com/oxy-hq/oxygen")
                 .support(
                     "- For support, please email robert@oxy.tech or contact us directly via Github."
                 )